Drumroll, fanfare, all that good stuff. It's code maintenance for v0.3.0 release :)

+++ configurationmanager_test.go
@@ -1,3 +1,12 @@
+/*
+	Digivance MVC Application Framework
+	Configuration Manage Unit Tests
+	Dan Mayor (dmayor@digivance.com)
+
+	This file defines the version 0.3.0 compatibility of configurationmanager.go functions. These functions are written
+	to demonstrate and test the intended use cases of the functions in configurationmanager.go
+*/
+
 package mvcapp_test
 
 import (
@@ -8,19 +17,41 @@ import (
 	"github.com/digivance/mvcapp"
 )
 
+// TestNewConfigurationManagerFromFile ensures that mvcapp.NewConfigurationManagerFromFile returns the expected value
 func TestNewConfigurationManagerFromFile(t *testing.T) {
 	configFile := mvcapp.GetApplicationPath() + "/testconfig.json"
+	errorConfigFile := mvcapp.GetApplicationPath() + "/failconfig.json"
+
 	configData := []byte("{\"HTTPPort\": 80,\"HTTPSPort\": 443,\"LogFilename\": \"./app.log\",\"LogLevel\": 4,\"TLSCertFile\": \"./mycert.crt\",\"TLSKeyFile\": \"./mycert.key\"}")
-	err := ioutil.WriteFile(configFile, configData, 0644)
+	errorConfigData := []byte("this is not json...")
+
+	config, err := mvcapp.NewConfigurationManagerFromFile(configFile)
+	if err == nil {
+		t.Error("Failed to fail on missing file")
+	}
+
+	err = ioutil.WriteFile(configFile, configData, 0644)
 	defer os.RemoveAll(configFile)
 
 	if err != nil {
 		t.Errorf("Failed to create new configuration manager json file: %s", err)
 	}
 
-	config, err := mvcapp.NewConfigurationManagerFromFile(configFile)
+	err = ioutil.WriteFile(errorConfigFile, errorConfigData, 0644)
+	defer os.RemoveAll(errorConfigFile)
+
 	if err != nil {
-		t.Errorf("Failed to create new configuration manager from file: %s", err)
+		t.Errorf("Failed to create new configuration manager json file: %s", err)
+	}
+
+	config, err = mvcapp.NewConfigurationManagerFromFile(errorConfigFile)
+	if err == nil {
+		t.Error("Failed to fail creating new configuration file from invalid json file")
+	}
+
+	config, err = mvcapp.NewConfigurationManagerFromFile(configFile)
+	if err != nil {
+		t.Errorf("Failed load config file: %s", err)
 	}
 
 	if config.HTTPPort != 80 || config.HTTPSPort != 443 || config.LogFilename != "./app.log" ||
@@ -29,6 +60,7 @@ func TestNewConfigurationManagerFromFile(t *testing.T) {
 	}
 }
 
+// TestConfigurationManager_SaveFile ensures that ConfigurationManager.SaveFile operates as expected
 func TestConfigurationManager_SaveFile(t *testing.T) {
 	config := &mvcapp.ConfigurationManager{
 		HTTPPort:    80,
@@ -39,7 +71,12 @@ func TestConfigurationManager_SaveFile(t *testing.T) {
 		TLSKeyFile:  "./mycert.key",
 	}
 
-	err := config.SaveFile("./testconfig.json")
+	err := config.SaveFile("?:\\*$%«╝╗")
+	if err == nil {
+		t.Error("Failed to fail to save to invalid filename")
+	}
+
+	err = config.SaveFile("./testconfig.json")
 	if err != nil {
 		t.Errorf("Failed to save configuration file: %s", err)
 	}
